Add files via upload
The company receives many loan applications on a daily basis and the loan is granted based on the applicant’s profile. Since the amount of applications is increasing, the company is facing issues in manually inspecting the loan application and approving/rejecting. To solve this issue and to automate the process the company is wanting you to build a machine learning model which can predict if the applicant is eligible for loan or not based on the historic data provided by the company.
A bank's profit or a loss depends to a large extent on loans i.e. whether the customers are paying back the loan or defaulting. By predicting the loan defaulters, the bank can reduce its Non- Performing Assets. This makes the study of this phenomenon very important